+++ src/datasources/data-frame/components/v1/DataFrameVariableQueryEditorV1.test.tsx
@@ -2,13 +2,13 @@ import { render, screen } from '@testing-library/react';
 import React from 'react';
 import { select } from 'react-select-event';
 import { setupDataSource } from 'test/fixtures';
-import { DataFrameDataSourceV1 } from '../../datasources/v1/DataFrameDataSourceV1';
 import { DataFrameQueryV1 } from '../../types';
 import { DataFrameVariableQueryEditorV1 } from './DataFrameVariableQueryEditorV1';
+import { DataFrameDataSource } from 'datasources/data-frame/DataFrameDataSource';
 
 const onChange = jest.fn();
 const onRunQuery = jest.fn();
-const [datasource] = setupDataSource(DataFrameDataSourceV1);
+const [datasource] = setupDataSource(DataFrameDataSource);
 
 test('renders with no data table selected', async () => {
   render(<DataFrameVariableQueryEditorV1 {...{ onChange, onRunQuery, datasource, query: '' as unknown as DataFrameQueryV1 }} />);