+++ src/datasources/data-frame/components/v2/DataFrameQueryEditorV2.tsx
@@ -1,27 +1,63 @@
-import React, { useCallback, useState } from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import { DataTableQueryBuilder } from "./query-builders/DataTableQueryBuilder";
-import { AutoSizeInput, Collapse, InlineField, InlineLabel, MultiSelect, RadioButtonGroup } from "@grafana/ui";
-import { DataFrameQueryV1, DataFrameQueryType, PropsV1 } from "../../types";
+import { AutoSizeInput, Collapse, Combobox, ComboboxOption, InlineField, InlineLabel, InlineSwitch, MultiCombobox, MultiSelect, RadioButtonGroup } from "@grafana/ui";
+import { DataFrameQueryV2, DataFrameQueryType, PropsV2, DataTableProjectionLabelLookup, DataTableProjectionType } from "../../types";
 import { enumToOptions, validateNumericInput } from "core/utils";
-import { TAKE_LIMIT } from 'datasources/data-frame/constants';
+import { decimationMethods, TAKE_LIMIT } from 'datasources/data-frame/constants';
+import { SelectableValue } from '@grafana/data';
+import { Workspace } from 'core/types';
+import { FloatingError } from 'core/errors';
 
-export const DataFrameQueryEditorV2: React.FC<PropsV1> = ({ query, onChange, onRunQuery, datasource }: PropsV1) => {
+export const DataFrameQueryEditorV2: React.FC<PropsV2> = ({ query, onChange, onRunQuery, datasource }: PropsV2) => {
     query = datasource.processQuery(query);
 
     const [isQueryConfigurationSectionOpen, setIsQueryConfigurationSectionOpen] = useState(true);
+    const [isColumnConfigurationSectionOpen, setIsColumnConfigurationSectionOpen] = useState(true);
+    const [isDecimationSettingsSectionOpen, setIsDecimationSettingsSectionOpen] = useState(true);
     const [recordCountInvalidMessage, setRecordCountInvalidMessage] = useState<string>('');
+    const [workspaces, setWorkspaces] = useState<Workspace[]>([]);
+
+    const getPropertiesOptions = (type: DataTableProjectionType) => Object.entries(DataTableProjectionLabelLookup)
+        .filter(([_, value]) => value.type === type)
+        .map(([key, value]) => ({ label: value.label, value: key })) as SelectableValue[];
+
+    const datatablePropertiesOptions = getPropertiesOptions(DataTableProjectionType.DataTable);
+    const columnPropertiesOptions = getPropertiesOptions(DataTableProjectionType.Column);
 
     const handleQueryChange = useCallback(
-        (query: DataFrameQueryV1, runQuery = true): void => {
+        (query: DataFrameQueryV2, runQuery = true): void => {
             onChange(query);
             if (runQuery) {
                 onRunQuery();
             }
         }, [onChange, onRunQuery]
     );
+
     const onQueryTypeChange = (queryType: DataFrameQueryType) => {
         handleQueryChange({ ...query, type: queryType }, false);
     };
+
+    const onColumnsChange = (columns: Array<ComboboxOption<string>>) => {
+        handleQueryChange({ ...query, columns: columns.map(i => i.value) }, false);
+    };
+
+    const onDecimationMethodChange = (option: ComboboxOption<string>) => {
+        handleQueryChange({ ...query, decimationMethod: option.value }, false);
+    };
+
+    const onUseTimeRangeChange = (event: React.FormEvent<HTMLInputElement>) => {
+        const value = event.currentTarget.checked;
+        handleQueryChange({ ...query, applyTimeFilters: value }, false);
+    };
+
+    useEffect(() => {
+        const loadWorkspaces = async () => {
+            const workspaces = await datasource.loadWorkspaces();
+            setWorkspaces(Array.from(workspaces.values()));
+        };
+
+        loadWorkspaces();
+    }, [datasource]);
 
     function validateTakeValue(value: number, TAKE_LIMIT: number) {
         if (isNaN(value) || value <= 0) {
@@ -66,6 +102,9 @@ export const DataFrameQueryEditorV2: React.FC<PropsV1> = ({ query, onChange, onR
                             placeholder={placeholders.datatableProperties}
                             width={valueFieldWidth}
                             onChange={(): void => { }}
+                            options={datatablePropertiesOptions}
+                            allowCustomValue={false}
+                            closeMenuOnSelect={false}
                         />
                     </InlineField>
                     <InlineField
@@ -77,6 +116,9 @@ export const DataFrameQueryEditorV2: React.FC<PropsV1> = ({ query, onChange, onR
                             placeholder={placeholders.columnProperties}
                             width={valueFieldWidth}
                             onChange={(): void => { }}
+                            options={columnPropertiesOptions}
+                            allowCustomValue={false}
+                            closeMenuOnSelect={false}
                         />
                     </InlineField>
                 </>
@@ -101,7 +143,10 @@ export const DataFrameQueryEditorV2: React.FC<PropsV1> = ({ query, onChange, onR
                         width: getValuesInPixels(valueFieldWidth),
                         marginBottom: getValuesInPixels(defaultMarginBottom)
                     }}>
-                        <DataTableQueryBuilder workspaces={[]} globalVariableOptions={[]} />
+                        <DataTableQueryBuilder
+                            workspaces={workspaces}
+                            globalVariableOptions={datasource.globalVariableOptions()}
+                        />
                     </div>
 
                     {query.type === DataFrameQueryType.Properties && (
@@ -124,7 +169,97 @@ export const DataFrameQueryEditorV2: React.FC<PropsV1> = ({ query, onChange, onR
                     )}
 
                 </Collapse>
-            </div>
+            </div >
+
+            {query.type === DataFrameQueryType.Data && (
+                <div
+                    style={{ width: getValuesInPixels(sectionWidth) }}
+                >
+                    <Collapse
+                        label={labels.columnConfigurations}
+                        isOpen={isColumnConfigurationSectionOpen}
+                        collapsible={true}
+                        onToggle={() => setIsColumnConfigurationSectionOpen(!isColumnConfigurationSectionOpen)}
+                    >
+                        <InlineField
+                            label={labels.columns}
+                            labelWidth={inlineLabelWidth}
+                            tooltip={tooltips.columns}
+                        >
+                            <MultiCombobox
+                                placeholder={placeholders.columns}
+                                width={inlineLabelWidth}
+                                onChange={onColumnsChange}
+                                options={[]}
+                                createCustomValue={false}
+                            />
+                        </InlineField>
+                        <InlineField
+                            label={labels.includeIndexColumns}
+                            labelWidth={inlineLabelWidth}
+                            tooltip={tooltips.includeIndexColumns}
+                        >
+                            <InlineSwitch
+                            />
+                        </InlineField>
+                        <InlineField
+                            label={labels.filterNulls}
+                            labelWidth={inlineLabelWidth}
+                            tooltip={tooltips.filterNulls}
+                        >
+                            <InlineSwitch
+                            />
+                        </InlineField>
+                    </Collapse>
+
+                    <Collapse
+                        label={labels.decimationSettings}
+                        isOpen={isDecimationSettingsSectionOpen}
+                        collapsible={true}
+                        onToggle={() => setIsDecimationSettingsSectionOpen(!isDecimationSettingsSectionOpen)}
+                    >
+                        <InlineField
+                            label={labels.decimationMethod}
+                            labelWidth={inlineLabelWidth}
+                            tooltip={tooltips.decimationMethod}
+                        >
+                            <Combobox
+                                width={inlineLabelWidth}
+                                onChange={onDecimationMethodChange}
+                                options={decimationMethods}
+                                createCustomValue={false}
+                            />
+                        </InlineField>
+                        <InlineField
+                            label={labels.xColumn}
+                            labelWidth={inlineLabelWidth}
+                            tooltip={tooltips.xColumn}
+                        >
+                            <Combobox
+                                placeholder={placeholders.xColumn}
+                                width={inlineLabelWidth}
+                                onChange={() => { }}
+                                options={[]}
+                                createCustomValue={false}
+                            />
+                        </InlineField>
+                        <InlineField
+                            label={labels.useTimeRange}
+                            labelWidth={inlineLabelWidth}
+                            tooltip={tooltips.useTimeRange}
+                        >
+                            <InlineSwitch
+                                onChange={onUseTimeRangeChange}
+                            />
+                        </InlineField>
+                    </Collapse>
+                </div >
+            )}
+            <FloatingError
+                message={datasource.errorTitle}
+                innerMessage={datasource.errorDescription}
+                severity="warning"
+            />
         </>
     );
 };
@@ -134,20 +269,37 @@ const labels = {
     datatableProperties: 'Data table properties',
     columnProperties: 'Column properties',
     queryConfigurations: 'Query configurations',
+    columnConfigurations: 'Column configurations',
+    decimationSettings: 'Decimation settings',
     queryByDatatableProperties: 'Query by data table properties',
+    columns: 'Columns',
+    filterNulls: 'Filter nulls',
+    includeIndexColumns: 'Include index columns',
+    decimationMethod: 'Decimation method',
+    xColumn: 'X-column',
+    useTimeRange: 'Use time range',
     take: 'Take',
 };
 const tooltips = {
     queryType: 'This field specifies the type for the query that searches the data tables. The query can retrieve row data or metadata.',
     queryByDatatableProperties: 'This optional field applies a filter to a query while searching the data tables.',
     take: 'This field sets the maximum number of records to return from the query.',
+    columns: 'Specifies the columns to include in the response data.',
+    filterNulls: `Specifies whether to filter out null and NaN values before decimating the data.`,
+    includeIndexColumns: 'Specifies whether to include index columns in the response data.',
     datatableProperties: 'This field specifies the data table properties to be queried.',
     columnProperties: 'This field specifies the column properties to be queried.',
+    decimationMethod: 'Specifies the method used to decimate the data.',
+    xColumn: 'Specifies the column to use for the x-axis when decimating the data.',
+    useTimeRange: `Specifies whether to query only for data within the dashboard time range if the
+                table index is a timestamp. Enable when interacting with your data on a graph.`,
 };
 const placeholders = {
     datatableProperties: 'Select data table properties to fetch',
     columnProperties: 'Select column properties to fetch',
-    take: 'Enter record count'
+    take: 'Enter record count',
+    columns: 'Select columns',
+    xColumn: 'Select x-column',
 };
 
 const errorMessages = {